Add wd_test() rule and define a test with it.
blob-test.js is copied directly from our internal test suite, but adapted to use the new `test()` entrypoint.

TODO: It's annoying that we need two files for a test. Perhaps the `.capnp` format could be extended with yet another way of writing strings that lets the JavaScript code be embedded in the capnp file in a way that is convenient to read and edit.